Fix: バックエンドの例外エラーをフロントに返さないように修正 (#72)
